get data with populate

+++ src/app/meals/meal.service.ts
@@ -7,22 +7,41 @@ import { Model } from 'mongoose';
 
 import { Meals } from './meal.entity';
 import { CreateMealDto } from './dto/create-meal.dto';
+import { Users } from '../users/user.entity';
+import { Nutrients } from '../nutrient/nutrient.entity';
 
 @Injectable()
 export class MealService {
   constructor(
     @InjectModel(Meals.name) private mealModel: Model<Meals>,
+    @InjectModel(Users.name) private userModel: Model<Users>,
   ) { }
 
-  async create(createMealDto: CreateMealDto): Promise<Meals> {
-    try {
-      const { ...mealData } = createMealDto;
-      const newMeal = new this.mealModel(mealData);
-      return await newMeal.save();
-    } catch (error) {
-      throw new HttpException(error.message, HttpStatus.BAD_REQUEST);
-    }
+ async create(createMealDto: CreateMealDto): Promise<Meals> {
+  const { name, date, totalCalories, mealType, notes, userId, nutrients } = createMealDto;
+
+  // Create new meal object
+  const newMeal = new this.mealModel({
+    name,
+    date,
+    totalCalories,
+    mealType,
+    notes,
+    user: userId
+  });
+
+  try {
+    const meal = await newMeal.save();
+    await this.userModel.findByIdAndUpdate(userId, {
+      $push: { meals: meal }
+    });
+
+    return meal;
+  } catch (error) {
+    throw new HttpException(error.message, HttpStatus.BAD_REQUEST);
   }
+}
+
 
   async findByname(name: string): Promise<Meals> {
     try {
@@ -73,14 +92,21 @@ export class MealService {
 
   async delete(id: string): Promise<{ message: string }> {
     try {
-      const deleted = await this.mealModel.findByIdAndDelete(id).exec();
-      if (!deleted) {
-        throw new HttpException("Meal not found", HttpStatus.BAD_REQUEST);
-      }
-      return { message: 'Meal deleted successfully' };
+        const meal = await this.mealModel.findById(id).exec();
+        if (!meal) {
+            throw new HttpException("Meal not found", HttpStatus.BAD_REQUEST);
+        }
+        await this.userModel.findByIdAndUpdate(meal.user, {
+            $pull: { meals: meal },
+        }).exec();
+        const deleted = await this.mealModel.findByIdAndDelete(id).exec();
+        if (!deleted) {
+            throw new HttpException('Meal not found', HttpStatus.BAD_REQUEST);
+        }
+        return { message: 'Meal deleted successfully' };
     } catch (error) {
-      throw new HttpException(error.message, HttpStatus.BAD_REQUEST);
+        throw new HttpException(error.message, HttpStatus.BAD_REQUEST);
     }
-  }
+}
 
 }
